requirements.txt
Flask-Script
Flask-SQLAlchemy
Flask-Login
WTForms
SQLAlchemy